feat: add filter buttons for OLTP and OLAP queries
This commit introduces two new buttons, "Only OLTP" and "Only OLAP", which allow users to filter the selected queries based on their type.  It also improves the query preview display to show the number of OLTP and OLAP queries and the total number of queries. Finally, it hides the previous result when the user clicks on the generate button.